# Provide direction argument to edge_diffs.