@hyanwong 2022-02-04
> Moves the implementation to Python resulting in a roughly 2X perf regression. I doubt this will be significant in practise, because people will usually be _doing_ something with the diffs, which will probably be more expensive. Nonetheless, it is a regression so we should be careful.

I'm fairly sure that we use the edge diffs a lot in tsdate, don't we @awohns ? Perhaps it's worth testing if this causes any perf problems there (I suspect not though, as we are doing some maths on each diff)