@jeromekelleher 2022-05-20
@benjeffery - should be push this in for 0.5.0?

@awohns - what's the simplest way of testing out the perf implications for tsdate?