@benjeffery 2022-05-20
If we are doing this (which seems the right thing to do I think) then yes it should be in 0.5.0.