@awohns 2022-05-26
Results of some perf testing on this:
`ts.edge_diffs()` is used in two places in tsdate. The more expensive usage is constructing the prior. Here are the runtimes for that operation on an inferred tree sequence from 1000 genomes chromosome 22. _**Current**_ is the current implementation of `edge_diffs()` and _**Direction**_ is the implementation in this PR:

Current: 6 mins 44 seconds. Direction: 7mins 41 seconds.