@jeromekelleher 2022-05-26
Hmm, so looks like most of the time for this code is actually spent constructing edge objects and the difference is really about the different code paths we take for that now. Some stats:

- C code path, yield None (i.e., don't create any edges): 0.98 seconds
- Python code path, yield None: 1.91 seconds
- C code path, yield lists of tuples: 1.19 seconds
- Python code path, yield lists of tuples: 4.05 seconds
- C code path, yield list of locally constructed Edge: 7.16 seconds (This is the current implementation in main)
- Python code yield list of Edge using self.edge():  13.66 (implementation Wilder tried out)
- Python code yielding list of locally constructed Edge: 10.55 seconds

So, I think there's room for quite a bit of improvement here and we did have a pretty big perf regression.

I wonder if edge metadata is really worth it here, and if we should do something like this:
```
# fully locked down dataclass
class DiffEdge:
     left: float
     right: float
     parent: int
     child: int
     id: int
    
     @property
     def metadata(self):
          raise ValueError("Metadata access not supported on edge diffs. Please use ``ts.edge(e.id).metadata`` instead")
```

I think if we can get back to the performance of yielding lists of tuples, we can switch the the Python code path and get rid of some C complexity, as this will be a perf improvement over what's currently in the library.

@benjeffery, any thoughts?