@benjeffery 2022-06-07
I've pushed an additional commit here which removes a function call in the `Edge` creation code path.

Here are the timings on my machine for 100k trees of 1.6M edges:

`main` 3s
This PR before commit (`ts.edge`): 7.12s
`Edge(*tuple)`: 4.8s
`Edge(*tuple) with `own_init`: 3.73s

So we're losing about 25% vs `main` with this python only code. I think that is acceptable and we should wrap this up by removing the C code and documenting the function if you agree?