@awohns 2022-06-15
Sorry, didn't see this until now @jeromekelleher.

So here's the timing for creating the tsdate prior of NYGC 1000 genomes chr22 (same file as before; 453,350 trees):
Current: 6 mins 29 seconds (389 seconds)
Direction: 7 mins 14 seconds (434 seconds)
So it's ~12% slower

And here's the perf for running `[naive_split_edges()](https://github.com/awohns/ldgm/blob/a152cb6db42658ab24d973757979af83a1c9c6d5/ldgm/bricks.py#L42)` on the first 30 MB of the same tree sequence (148,437 trees):
Current: 8 mins 57 seconds
Direction: 8 mins 2 seconds (482 seconds)
Which I guess means that `edge_diffs()` is not really a contributing factor to the overall runtime.