@awohns 2022-06-16
Thanks for seeking my feedback @jeromekelleher and for input @benjeffery! 
I think that 10% is acceptable from my perspective, especially given that `edge_diffs` is not a dominant factor in the runtime of either `tsdate` or `ldgm`. So I'm happy for this to be merged!