@jeromekelleher 2022-06-20
This also still leaves in the C infrastructure as dead code. Will pull out now.