@benjeffery 2022-06-20
> whoops, took off the AUTOMERGE label there @benjeffery as I'd left some embarassing intermediate commits.Will squash those now.

Gah, sorry. I didn't think properly about the C and docs.